requirements.txt
requests==2.33.1
beautifulsoup4==4.14.3
python-telegram-bot==22.7
APScheduler==3.11.2
pytz==2026.1.post1
playwright==1.48.0
gunicorn==21.2.0